chore: release 1.1.0, docs sync (JP), remove legacy node.py shim
