+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+### Changed
+### Fixed
+
+## [1.1.0] - 2025-09-24
+Note: 1.0.0 was the first public registry release; this minor release formalizes post‑1.0 refactor cleanup (shim removal) and new UI/node enhancements.
+
+### Added
 - Node: `Show Any (Any to String)` — accepts any input, converts to STRING, displays on canvas; supports batching.
 - Frontend: `web/show_text_unimeta.js` extended to handle `ShowAny|unimeta` and `ShowText|unimeta` with robust payload parsing.
 - Frontend: Dynamic textarea sizing and node recompute to reduce overlap at small zoom levels.
@@ -13,6 +20,11 @@ All notable changes to this project will be documented in this file.
 ### Changed
 - Frontend separation: `web/show_text.js` now targets base `ShowText` only to avoid double initialization with UniMeta variants.
 - Improved truncation suffix documentation and test expectations for `_safe_to_str`.
+- Removed legacy compatibility shim `saveimage_unimeta/nodes/node.py`; direct imports now required:
+	- `SaveImageWithMetaDataUniversal` → `saveimage_unimeta.nodes.save_image`
+	- `SaveCustomMetadataRules` → `saveimage_unimeta.nodes.rules_writer`
+	- `MetadataRuleScanner` → `saveimage_unimeta.nodes.scanner`
+	- Centralized EXIF test monkeypatch target: `saveimage_unimeta.piexif_alias.piexif`.
 
 ### Fixed
 - Prevented double widget injection causing textarea overlap in UniMeta nodes.
