+++ saveimage_unimeta/nodes/__init__.py
@@ -1,11 +1,9 @@
-from .node import (  # noqa: N999 - package path inherited from external naming constraints
-    CreateExtraMetaDataUniversal,
-    SaveCustomMetadataRules,
-    SaveGeneratedUserRules,
-    SaveImageWithMetaDataUniversal,
-    ShowGeneratedUserRules,
-    MetadataRuleScanner,
-)
+from .save_image import SaveImageWithMetaDataUniversal  # extracted from node.py
+from .extra_metadata import CreateExtraMetaDataUniversal  # extracted from node.py
+from .rules_view import ShowGeneratedUserRules  # extracted from node.py
+from .rules_save import SaveGeneratedUserRules  # extracted from node.py
+from .scanner import MetadataRuleScanner  # extracted from node.py
+from .rules_writer import SaveCustomMetadataRules  # moved out of legacy node.py
 from .show_text import ShowText  # local unimeta variant (separate file for clarity)
 from .show_any import ShowAnyToString  # new any->string display node
 from ..defs import set_forced_include
@@ -93,8 +91,12 @@ class MetadataForceInclude:
 
 __all__ = [
     "SaveImageWithMetaDataUniversal",
+    "CreateExtraMetaDataUniversal",
     "MetadataForceInclude",
     "MetadataRuleScanner",
+    "ShowGeneratedUserRules",
+    "SaveGeneratedUserRules",
+    "SaveCustomMetadataRules",
 ]
 
 NODE_CLASS_MAPPINGS = {
